==> PCode/2_OneHot.ipynb <==
# 2. One Hot



Seeing as the data there are many categorical variables, we proceed to one-hot the data

**import packages**

import numpy as np
import pandas as pd
import pickle
import os
from sklearn.preprocessing import OneHotEncoder

from matplotlib import pyplot as plt
%matplotlib inline

**set directories**

print(os.getcwd())
dirRawData = "../RawData/"
dirPData = "../PData/"
dirPOutput = "../POutput/"

**open data from pickle**

f_name = dirPData + '01_df_all.pickle'

with (open(f_name, "rb")) as f:
    dict_ = pickle.load(f)

    
vars_ind_num = dict_['vars_ind_num']
vars_ind_categorical = dict_['vars_ind_categorical']
df_all = dict_['df_all']

del f_name, dict_

vars_all = df_all.columns.values
var_dep = ['target']

vars_ind = [var for var in vars_all if var not in (var_dep)]

vars_ind_value = df_all[vars_ind] #values of all the independent variables
vars_ind_value.shape

**loop to one-hot variables**

vars_ind_onehot = []
df_all_onehot = df_all.copy()
for col in vars_ind_categorical:
    print(col)
    enc_ = OneHotEncoder()
    enc_.fit(df_all[[col]])
    oh = enc_.transform(df_all[[col]]).toarray()
    idx_mostFreq = np.argmax(oh.sum(axis=0))
    oh = np.delete(oh, idx_mostFreq, axis=1)#delete most frequent unique value for particular categorical variable
    oh_names = enc_.categories_[0]
    oh_names = np.delete(oh_names, idx_mostFreq)
    oh_names = col + '_X_' + oh_names
    df_oh = pd.DataFrame(data = oh, columns = oh_names)
    df_all_onehot = pd.concat([df_all_onehot, df_oh], axis = 1, sort = False)
    del df_all_onehot[col]
    vars_ind_onehot.extend(oh_names)

df_all_onehot.dtypes

#split out test and train data
df_train_onehot = df_all_onehot[df_all_onehot['target'].isna()==False]
df_test_onehot = df_all_onehot[df_all_onehot['target'].isna()==True]

#split train data into 10 folds (0 to 9)
rng = np.random.RandomState(2018)
fold = rng.randint(0,10,df_train_onehot.shape[0])
df_train_onehot.loc[:,'fold']= fold

#let df_test_onehot have fold column with value 10
df_test_onehot.loc[:,'fold']= 10

#add back train and test dataframe to form a new df_all_onehot
df_all_onehot = pd.concat([df_train_onehot, df_test_onehot],ignore_index = True, sort =False)

#store dataset
dict_ = {'df_all': df_all,
         'df_all_onehot': df_all_onehot}

f_name = dirPData + '02_df.pickle'

with open(f_name, "wb") as f:
    pickle.dump(dict_, f)
    
del f_name, dict_

dict_ = {'vars_ind_num': vars_ind_num, 
         'var_dep': var_dep,
         'vars_ind_categorical': vars_ind_categorical,
         'vars_ind_onehot': vars_ind_onehot
         }


f_name = dirPData + '02_vars.pickle'

with open(f_name, "wb") as f:
    pickle.dump(dict_, f)
    
del f_name, dict_
