==> PCode/10_Ensemble_Stacking.ipynb <==
## 10. Ensembling

#### Import libraries

import os
import numpy as np
import pandas as pd
import pickle

from sklearn.preprocessing import StandardScaler # to standardize features 

import matplotlib.pyplot as plt
%matplotlib inline

print(os.getcwd())
dirRawData = "../RawData/"
dirPData = "../PData/"
dirPOutput = "../POutput/"

#### Load the data

# load df_all
f_name = dirPData + '02_df.pickle'
with open(f_name, "rb") as f:
    dict_ = pickle.load(f)
    
df_all_onehot = dict_['df_all_onehot']

del f_name, dict_

# load the dict_ that was saved to pickle
#dict_ = {'vars_ind': vars_ind, 'idx_design': idx_design, 'idx_test': idx_test}
f_name = dirPData + '02_vars.pickle'
with open(f_name, "rb") as f:
    dict_ = pickle.load(f)
    
var_dep = dict_['var_dep']
vars_ind_num = dict_['vars_ind_num']
vars_ind_categorical = dict_['vars_ind_categorical']
vars_ind_onehot = dict_['vars_ind_onehot']

del f_name, dict_

#### Preparing the data

#Create indexes for the train, validatiom, design and test set
idx_train  = np.where(df_all_onehot['fold'].isin(np.arange(0,8)))[0] 
idx_val    = np.where(df_all_onehot['fold'].isin([8,9]))[0]
idx_design = np.where(df_all_onehot['fold'].isin(np.arange(0,10)))[0]
idx_test   = np.where(df_all_onehot['fold'].isin([10]))[0]

#Check if the validation set and the train set is the same as the design set
len(idx_val)+len(idx_train) == len(idx_design)

#Store the length of the indexes 
n_train = len(idx_train)
n_val = len(idx_val)
n_test = len(idx_test)

#store all the independent variables into one
vars_ind = vars_ind_num + vars_ind_onehot

X = df_all_onehot[vars_ind].values #extract the values

# Standardise (each feature to have mean:0 and stdev:1)
X_train  = X[idx_train, :]

standardScaler_ = StandardScaler()
standardScaler_.fit(X_train)

# Now standardise all the data
X = standardScaler_.transform(X)

X_design = X[idx_design, :]
X_train  = X[idx_train, :]
X_val    = X[idx_val, :]
X_test   = X[idx_test, :]

y_design = df_all_onehot[var_dep].iloc[idx_design].copy().values.ravel()
y_train  = df_all_onehot[var_dep].iloc[idx_train].copy().values.ravel()
y_val    = df_all_onehot[var_dep].iloc[idx_val].copy().values.ravel()
y_test   = df_all_onehot[var_dep].iloc[idx_test].copy().values.ravel()

# Load the models

###### 1. Xgboost

Kaggle Score: 0.87807

f_name = dirPData + 'xgb.pickle'
with open(f_name, "rb") as f:
    dict_ = pickle.load(f)  
xgb1_train = dict_['xgbopt_prob']
xgb1_test = dict_['xgbopt_prob_test']
    
del f_name, dict_

xgb1_train = xgb1_train[:,1]
xgb1_test = xgb1_test[:,1]

###### 2. Logistic Regression

Kaggle score: 0.86802

f_name = dirPData + 'logreg.pickle'
with open(f_name, "rb") as f:
    dict_ = pickle.load(f)  
log_reg_design = dict_['lropt_prob']
log_reg_test = dict_['lropt_prob_test']
    
del f_name, dict_

lr_train = log_reg_design[:,1]
lr_test = log_reg_test[:,1]

###### 3. Elastic Net

Kaggle score: 0.85632

f_name = dirPData + 'EN.pickle'
with open(f_name, "rb") as f:
    dict_ = pickle.load(f)  
EN_train = dict_['ENopt_prob']
EN_test = dict_['ENopt_prob_test']
    
del f_name, dict_

EN_train = EN_train[:,1]
EN_test = EN_test[:,1]

**4. Decision Tree**

Kaggle Score: 0.78208

f_name = dirPData + 'dt.pickle'
with open(f_name, "rb") as f:
    dict_ = pickle.load(f)  
dt_train = dict_['dt_prob_design']
dt_test = dict_['dt_prob_test']
    
del f_name, dict_

#obtain the probabilities of crash
dt_train = dt_train[:,1]
dt_test = dt_test[:,1]

**5. KNN**

Kaggle Score:  0.67657

f_name = dirPData + 'knn.pickle'
with open(f_name, "rb") as f:
    dict_ = pickle.load(f)  
knn_train = dict_['knn_prob_design']
knn_test = dict_['knn_prob_test']
    
del f_name, dict_

knn_train = knn_train[:,1]
knn_test = knn_test[:,1]
knn_train.shape

###### 6. lightGBM model

Kaggle Score: 0.87285

f_name = dirPData + 'lgbm.pickle'
with open(f_name, "rb") as f:
    dict_ = pickle.load(f) 
lgbm_train = dict_['lgbm_prob_design']
lgbm_test = dict_['lgbm_prob_test']
    
del f_name, dict_

lgbm_train.shape

###### 7. Random Forest

Kaggle Score: 0.85673

f_name = dirPData + 'rf.pickle'
with open(f_name, "rb") as f:
    dict_ = pickle.load(f) 
rf_train = dict_['rf_prob_design']
rf_test = dict_['rf_prob_test']
    
del f_name, dict_

rf_train.shape

**Combine train and test ensemble data into numpy arrays**

#combine design data into array
combined = np.concatenate((EN_train.reshape(-1,1),lr_train.reshape(-1,1),knn_train.reshape(-1,1),
                           xgb1_train.reshape(-1,1), lgbm_train.reshape(-1,1),
                           rf_train.reshape(-1,1),dt_train.reshape(-1,1)),axis=1)

combined.shape

#combine test data into array
combined_test = np.concatenate((EN_test.reshape(-1,1),lr_test.reshape(-1,1),knn_test.reshape(-1,1),
                                xgb1_test.reshape(-1,1), lgbm_test.reshape(-1,1),
                                rf_test.reshape(-1,1),dt_test.reshape(-1,1)),axis=1)

**XGBoost on stacked ensemble data**

Stacking is an ensemble learning technique that uses predictions from multiple models (for example decision tree, knn or svm) to build a new model. This model is used for making predictions on the test set. 

#instantiate and fit model to run on stack
import xgboost as xgb
xgb_ = xgb.XGBClassifier()
xgb_.fit(X = combined, y = np.ravel(y_design)) # fit model to train data with default settings

from sklearn.model_selection import GridSearchCV
#tune all 3 parameters simultaneously
min_child_weight = [1,3,5,7]
gamma = [0,1,2,3]
max_depth=[1,2,3,4,5,6]
param_grid = dict(gamma = gamma, min_child_weight = min_child_weight, max_depth=max_depth)
xgb_ = xgb.XGBClassifier()
grid = GridSearchCV(estimator=xgb_, param_grid=param_grid, cv = 10, n_jobs=-1,scoring='roc_auc')
#cv = number of folds for validation
grid_result = grid.fit(combined,np.ravel(y_design))
print("Best: %f using %s" % (grid_result.best_score_, grid_result.best_params_)) # Summarize results

ensemble_proba2 = grid.predict_proba(combined_test)

This is an almost perfect AUC score however Kaggle Score was only 0.78648. This could be due to overfitting the models.

**Save data**

import csv
f_name = dirPOutput + 'ensemble_pred2.csv'

df_test = pd.read_csv(dirRawData + 'test.csv')

with open(f_name, 'w',newline='') as csvfile:
    writer=csv.writer(csvfile,delimiter=',')
    writer.writerow(["id", "target"])
    writer.writerows(zip(df_test[df_test.columns[0]], ensemble_proba2[:,1]))

**Logistic regression on ensemble data**

from sklearn.model_selection import GridSearchCV
from sklearn.linear_model import LogisticRegression
C=[0.05,0.1,1.0,1.5,2.0]
penalty = ["l1","l2"]
param_grid = dict(penalty = penalty,C=C)
lr_= LogisticRegression(solver ='liblinear')
grid = GridSearchCV(estimator=lr_, param_grid=param_grid, cv = 10, n_jobs=-1, scoring = 'roc_auc')
#cv = number of folds for validation

#fit using best parameters
grid_result = grid.fit(combined,np.ravel(y_design))
print("Best: %f using %s" % (grid_result.best_score_, grid_result.best_params_)) # Summarize results

ensemble_proba3 = grid.predict_proba(combined_test)

#save prediction as csv
import csv
f_name = dirPOutput + 'ensemble_pred3.csv'

df_test = pd.read_csv(dirRawData + 'test.csv')

with open(f_name, 'w',newline='') as csvfile:
    writer=csv.writer(csvfile,delimiter=',')
    writer.writerow(["id", "target"])
    writer.writerows(zip(df_test[df_test.columns[0]], ensemble_proba3[:,1]))

The AUC score is still high but Kaggle Score was only 0.7702. This could be due to overfitting the models.

**Plot correlation matrix**

#save all as dataframe

base_predictions_train = pd.DataFrame( {'RandomForest': rf_train.ravel(),
                                        'DecisionTree': dt_train.ravel(),
                                        'LogisticReg': lr_train.ravel(),
                                        'XGB': xgb1_train.ravel(),
                                        'LGBM': lgbm_train.ravel(),
                                        'ElasticNet': EN_train.ravel(),
                                        'KNN': knn_train.ravel()                                      
    })

base_predictions_train.head()




#print corelation matrix
corr = base_predictions_train.corr()
corr.style.background_gradient(cmap='coolwarm')

It seems that Random Forest is highly correlated with XGB and LGBM, thus we should only select one out of the three. ElasticNet and LogisticReg are also highly correlated, which is no surprise. In the end the only models would then be decision tree, XGB, LogisticReg and KNN. Alternately we could take the average of the correlated model predictions.

#try running models without the correlated pairs

#combine design data into array
combined = np.concatenate((lr_train.reshape(-1,1),knn_train.reshape(-1,1),
                           xgb1_train.reshape(-1,1),dt_train.reshape(-1,1)),axis=1)

#combine test data into array
combined_test = np.concatenate((lr_test.reshape(-1,1),knn_test.reshape(-1,1),
                                xgb1_test.reshape(-1,1),dt_test.reshape(-1,1)),axis=1)

min_child_weight = [1,3,5,7]
gamma = [0,1,2,3]
max_depth=[1,2,3,4,5,6]
param_grid = dict(gamma = gamma, min_child_weight = min_child_weight, max_depth=max_depth)
xgb_ = xgb.XGBClassifier()
grid = GridSearchCV(estimator=xgb_, param_grid=param_grid, cv = 10, n_jobs=-1,scoring='roc_auc')
#cv = number of folds for validation
grid_result = grid.fit(combined,np.ravel(y_design))
print("Best: %f using %s" % (grid_result.best_score_, grid_result.best_params_)) # Summarize results

ensemble_proba4=grid.predict_proba(combined_test)

#save prediction as csv
import csv
f_name = dirPOutput + 'ensemble_pred4.csv'

df_test = pd.read_csv(dirRawData + 'test.csv')

with open(f_name, 'w',newline='') as csvfile:
    writer=csv.writer(csvfile,delimiter=',')
    writer.writerow(["id", "target"])
    writer.writerows(zip(df_test[df_test.columns[0]], ensemble_proba4[:,1]))

Sadly, the Kaggle Score still was only 0.77169..
